Revert "Plan 9 from Bell Labs Fourth Edition 2012-07-31"
This reverts commit 48207d97b4e803eefbe4064a810e27dec2c3d9bb.

+++ sys/src/libmach/5db.c
@@ -298,7 +298,7 @@ plocal(Instr *i)
  * Print value v as name[+offset]
  */
 static int
-gsymoff(char *buf, int n, ulong v, int space)
+gsymoff(char *buf, int n, long v, int space)
 {
 	Symbol s;
 	int r;
@@ -1011,7 +1011,7 @@ format(char *mnemonic, Instr *i, char *f)
 
 		case 'b':
 			i->curr += symoff(i->curr, i->end-i->curr,
-				(ulong)i->imm, CTEXT);
+				i->imm, CTEXT);
 			break;
 
 		case 'g':
